pyproject.toml
[tool.poetry]
name = "mizu-node"
version = "0.1.0"
description = ""
authors = ["shudong <dongs2011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
redis = "^5.1.1"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
requests = "^2.32.3"
fastapi-cors = "^0.0.6"
prometheus-client = "^0.21.0"
psycopg2-binary = "^2.9.10"

sqlalchemy = "^2.0.36"
aioboto3 = "^13.2.0"
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^4.0.1"
black = "^24.10.0"
isort = "^5.13.2"
cryptography = "^43.0.3"
httpx = "^0.27.2"
pytest-mock = "^3.14.0"
warcio = "^1.7.4"
aiohttp = "^3.10.10"
freezegun = "^1.5.1"
testing-postgresql = "^1.3.0"
pymongo = "^4.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "mizu_node.node_server:start_dev"
start = "mizu_node.node_server:start"
publish_pow = "publisher.pow:start"
publish_reward = "publisher.reward:start"
publish_batch_classify = "publisher.batch_classify:start"

backup = "scripts.backup:main"
client = "scripts.client:main"
import = "scripts.importer:main"